pyproject.toml
[tool.poetry]
name = "Emotion-model"
version = "0.1.0"
description = ""
authors = [
    "Maria Yadryshnikova <myadryshnikova@tourmalinecore.com>",
]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.10"
boto3 = "^1.21.42"
boto3-stubs-lite = "^1.26.8"
opencv-python-headless = "^4.6.0.66"
pillow = "^9.3.0"
scikit-learn = "^1.1.3"
dvc = {extras = ["gdrive"], version = "^2.34.2"}
pydrive2 = "^1.15.0"
pyinstaller = "^5.6.2"
transformers = "^4.24.0"
pika = "^1.3.1"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.10.0"
pylint = "^2.15.5"
pylint-quotes = "^0.2.3"
pylint-flask = "^0.6"
pylint-flask-sqlalchemy = "^0.2.0"
pylint-exit = "^1.2.0"
